==> 01 Predicting Bike Rentals - Linear reg vs XGBoost.ipynb <==
# Data Wrangling

import xgboost

from pandas import MultiIndex, Int64Index

import platform; print(platform.platform())

import sys; print("Python", sys.version)

import numpy; print("NumPy", numpy.__version__)

import scipy; print("SciPy", scipy.__version__)

import sklearn; print("Scikit-Learn", sklearn.__version__)

import xgboost; print("XGBoost", xgboost.__version__)

import pandas as pd

### Load data (file was saved in local directory)

df_bikes = pd.read_csv('bike_rentals.csv')

### Understanding the data

df_bikes.head()

df_bikes.describe()

df_bikes.info()

### Correcting null values

#Find the number of null values:
df_bikes.isna().sum().sum()

#Display null values:
df_bikes[df_bikes.isna().any(axis=1)]

#### Replacing column values with mean/median

# windspeed column has a good amount of null values. Since mean can be vulnerable to outliers, will replace with median values.

df_bikes['windspeed'].fillna((df_bikes['windspeed'].median()), inplace=True)

#checking the column values:
df_bikes.iloc[[56, 81, 128, 298,528]]

#### Replacing column values with mean/median with groupby

Humidity (hum) can easily be dependent on the changes from different seasons.It would make more sense to replace null values of hum with medians of the different seasons. 

df_bikes.groupby(['season']).median()

# the null values in hum column are corrected by the values obtained from groupby statement. It requires "transform" method:
df_bikes['hum'] = df_bikes['hum'].fillna(df_bikes.groupby('season')['hum'].transform('median'))

#checking the column values:
df_bikes.iloc[[129, 213, 388]]

#### Obtaining the median/mean from specific rows

For some variables, it'd make more sense to replace null values with the mean of specific rows.
For example, when correcting temperature, rather than using all historical records, taking the mean temperature of the day before and the day after should give a good estimate.

df_bikes[df_bikes['temp'].isna()]

#Sum temperatures in specific rows and take mean.
mean_temp = (df_bikes.iloc[700]['temp'] + df_bikes.iloc[702]['temp'])/2

mean_atemp = (df_bikes.iloc[700]['atemp'] + df_bikes.iloc[702]['atemp'])/2

# Replacing null with mean values
df_bikes['temp'].fillna((mean_temp), inplace=True)

df_bikes['atemp'].fillna((mean_atemp), inplace=True)

df_bikes.iloc[[701]]

#### Extrapolating dates 

"dteday" column is defined as object (see .info()), commonly represented as string.
However, date objects such as years and months must be extrapolated from datetime types..

#convert the 'dteday'column to 'datetime' type by using to_datetime method:
df_bikes['dteday'] = pd.to_datetime(df_bikes['dteday'],infer_datetime_format=True)

#import datetime library
import datetime as dt

#convert the 'mnth' column to the correct months extrapolated from the 'dteday' column. 
df_bikes['mnth'] = df_bikes['dteday'].dt.month

df_bikes.tail()

#use the .loc method to fill in the correct value for 'yr'
# data is normalized, thus we observe yr values as "1.0". Normalized data is more efficient for weighting purposes.. 
df_bikes.loc[730, 'yr'] = 1.0

#### Deleting non-numerical columns

dteday is non-numerical column and also it's redundant. This column may be deleted.

df_bikes = df_bikes.drop('dteday', axis=1)

## Regression Analysis - Bikes Rentals

'cnt' is the number of bike rentals in a given day. I will predict this column in this project.
Opposed to other transport services such as bus or subway, the duration of travel, departure and arrival position is explicitly recorded in bike sharing systems. This feature turns the system into a virtual sensor network that can be used for sensing mobility in the city. Hence, it is expected that most of important events in the city could be detected via monitoring these data.

# df_bikes['cnt'] is the sum of df_bikes['casual'] and df_bikes['registered'].So those columns will not be needed for prediction.
#Drop the 'casual' and 'registered' columns using the .drop method:
df_bikes = df_bikes.drop(['casual', 'registered'], axis=1)

#### Saving data for future use:

df_bikes.to_csv('bike_rentals_cleaned.csv', index=False)

#### Predictor and target columns

X = df_bikes.iloc[:,:-1] #The first colon, :, means that all rows are included. After the comma, :-1 means start at the first column and go all the way to the last column without including it.
y = df_bikes.iloc[:,-1] # -1 takes the last column only.

from sklearn.model_selection import train_test_split

from sklearn.linear_model import LinearRegression

# train-test-split

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=2)
#random_state: choosing the seed of a pseudo-random number generator to ensure reproducible results

### Modeling Linear Regression

# initialize machine learnin model
lin_reg = LinearRegression()

# Fit the model on training set
lin_reg.fit(X_train, y_train)

# Make predictions of the test set
y_pred = lin_reg.predict(X_test)

#Compare the predictions with the test set. Scoring the model requires a basis of comparison. The standard for linear regression is the root mean squared error (RMSE).
from sklearn.metrics import mean_squared_error

import numpy as np

mse = mean_squared_error(y_test, y_pred)

rmse = np.sqrt(mse)

# print results:
print("RMSE: %0.2f" % (rmse))

# Is an error of 898 rentals good or bad ?
# Check the expected range of rentals per day.
df_bikes['cnt'].describe()

### With a range of 22 to 8714, a mean of 4504, and a standard deviation of 1937, 
# an RMSE of 898 isn't bad, but it's not great either.

### XGBRegressor

XGBRegressor can be used on any regression dataset. I will use it to compare results on the bike rentals dataset with linear regression.

from xgboost import XGBRegressor

# Initialize a machine learning model:
xg_reg = XGBRegressor()

#Fit the model on the training set:
xg_reg.fit(X_train, y_train)

#Make predictions for the test set:
y_pred = xg_reg.predict(X_test)

#Compare the predictions with the test set:
mse = mean_squared_error(y_test, y_pred)

rmse = np.sqrt(mse)

#Print results:
print("RMSE: %0.2f" % (rmse))

#### XGBRegressor performs better than Linear reg.

### Cross-validation

One test score is not reliable because splitting the data into different training and test sets would give different results. In effect, splitting the data into a training set and a test set is arbitrary, and a different random_state will give a different RMSE.

One way to address the score discrepancies between different splits is k-fold cross-validation. The idea is to split the data multiple times into different training sets and test sets, and then to take the mean of the scores. The number of splits, called folds, is denoted by k. It's standard to use k = 3, 4, 5, or 10 splits.
There is going to be overlap in the training sets, but not the test sets. At the end, there will be k different scores evaluating the model against k different test sets. Taking the mean score of the k folds gives a more reliable score than any single fold.

#### Cross-validation with linear regression

from sklearn.model_selection import cross_val_score

#Initialize a machine learning model:
model = LinearRegression()

# Implement cross_val_score with the model, X, y, scoring='neg_mean_squared_error', and the number of folds, cv=10, as input:
scores = cross_val_score(model, X, y, scoring='neg_mean_squared_error', cv=10)

# Find the RMSE by taking the square root of the negative scores:
rmse = np.sqrt(-scores)

# Display the results:
print('Reg rmse:', np.round(rmse, 2))

print('RMSE mean: %0.2f' % (rmse.mean()))

#Using cross-validation is always recommended for a better estimate of the score.

#### Cross-validation with XGBoost

# Initialize a machine learning model:
model = XGBRegressor()

# Implement cross_val_score with the model, X, y, scoring, and the number of folds, cv, as input:
scores = cross_val_score(model, X, y, scoring='neg_mean_squared_error', cv=10)

# Find the RMSE by taking the square root of the negative scores:
rmse = np.sqrt(-scores)

# Print the results:
print('Reg rmse:', np.round(rmse, 2))

print('RMSE mean: %0.2f' % (rmse.mean()))

###### XGBRegressor wins again, performing 10% better than linear regression.
