==> 02 Predicting Personal Income - Classification vs XGBoost.ipynb <==
### Data Wrangling

import xgboost

from pandas import MultiIndex, Int64Index

import platform; print(platform.platform())

import sys; print("Python", sys.version)

import numpy; print("NumPy", numpy.__version__)

import scipy; print("SciPy", scipy.__version__)

import sklearn; print("Scikit-Learn", sklearn.__version__)

import xgboost; print("XGBoost", xgboost.__version__)

import pandas as pd

#### Load data (file was saved in local directory)

df_census = pd.read_csv('adult.data')

#### Understanding the data

df_census.head()

df_census = pd.read_csv('adult.data', header=None)
df_census.head()

# Column names can be renamed as follows:
df_census.columns=['age', 'workclass', 'fnlwgt', 'education', 'education-num', 'marital-status', 'occupation', 'relationship', 'race', 'sex', 'capital-gain', 'capital-loss', 'hours-per-week', 'native-country', 'income']

df_census.head()

#### Check null values

df_census.info()

# There are no null values.

#### non-numerical columns
All columns with "object" type need to be transformed into numerical columns.
Pandas 'get_dummies' method takes the non-numerical unique values of every column and converts them into their own column, with 1 indicating presence and 0 indicating absence. 

# we can eliminate "education" column since it's a duplicate of education_num:
df_census = df_census.drop(['education'], axis=1)

# use pd.get_dummies to transform the non-numerical columns into numerical columns:
df_census = pd.get_dummies(df_census)

df_census.head()

#### Predictor and target columns

#The target column is whether or not someone makes 50K. Two columns, df_census['income_<=50K'] and df_census['income_>50K'], are used to determine whether someone makes 50K. 
#Since either column will work, we delete df_census['income_ <=50K']:

df_census = df_census.drop('income_ <=50K', axis=1)

X = df_census.iloc[:,:-1]
y = df_census.iloc[:,-1]

##### Saving data for future use:
df_census.to_csv('census_cleaned.csv', index=False)

### Logistic Regression

# import logistic regression:
from sklearn.linear_model import LogisticRegression

#### The cross-validation function

from sklearn.metrics import mean_squared_error

import numpy as np

from sklearn.model_selection import cross_val_score

def cross_val(classifier, num_splits=10):    
    model = classifier     
    scores = cross_val_score(model, X, y, cv=num_splits)    
    print('Accuracy:', np.round(scores, 2))    
    print('Accuracy mean: %0.2f' % (scores.mean()))

cross_val(LogisticRegression())

### The XGBoost classifier

#import classifier of XGBoost:
from xgboost import XGBClassifier

# Since there are 94 columns, and XGBoost is an ensemble method, meaning that it combines many models for each run, each of which includes 10 splits, we are going to limit n_estimators, the number of models, to 5.
cross_val(XGBClassifier(n_estimators=5))

#Accuracy: [0.85 0.86 0.87 0.85 0.86 0.86 0.86 0.87 0.86 0.86]
#Accuracy mean: 0.86

# XGBoost is the winner again..
