==> 05 Gradient boosting versus XGBoost - Discovering Exoplanets exercise.ipynb <==
The general idea behind boosting algorithms is to transform weak learners into strong learners. A weak learner is hardly better than random guessing. But there is a purpose behind the weak start. Building on this general idea, boosting works by focusing on iterative error correction, not by establishing a strong baseline model. If the base model is too strong, the learning process is necessarily limited, thereby undermining the general strategy behind boosting models.

Weak learners are transformed into strong learners through hundreds of iterations. Boosting has been one of the best general machine learning strategies in terms of producing optimal results. 

### Gradient Boosting

While gradient boosting also adjusts based on incorrect predictions (similar to AdaBoost), it takes this idea one step further: gradient boosting fits each new tree entirely based on the errors of the previous tree's predictions. That is, for each new tree, gradient boosting looks at the mistakes and then builds a new tree completely around these mistakes. The new tree doesn't care about the predictions that are already correct.
Gradient boosting computes the residuals of each tree's predictions and sums all the residuals to score the model.
A gradient boosting model is built from scratch by training new trees on the errors of the previous trees. 
We continue with the bike rentals dataset to compare new models with the previous models:

import pandas as pd
import numpy as np
import warnings
warnings.filterwarnings('ignore')

# Load the bike rental dataset:
df_bikes = pd.read_csv('bike_rentals_cleaned.csv')
df_bikes.head()

#### Building a gradient boosting model from scratch

# split the data into X and y. Then, split X and y into training and test sets:
X_bikes = df_bikes.iloc[:,:-1]
y_bikes = df_bikes.iloc[:,-1]
from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X_bikes, y_bikes, random_state=2)

# Fit the data to the decision tree:  The initial decision tree, called a base learner, should not be fine-tuned for accuracy. 
# We want a model that focuses on learning from errors, not a model that relies heavily on the base learner.
# Initialize a decision tree with max_depth=2 and fit it on the training set as tree_1, since it's the first tree in our ensemble:
from sklearn.tree import DecisionTreeRegressor
tree_1 = DecisionTreeRegressor(max_depth=2, random_state=2)
tree_1.fit(X_train, y_train)

#Make predictions with the training set: Instead of making predictions with the test set, predictions in gradient boosting are initially made with the training set. 
#To compute the residuals, we need to compare the predictions while still in the training phase. The test phase of the model build comes at the end, after all the trees have been constructed. 
#The predictions of the training set for the first round are obtained by adding the predict method to tree_1 with X_train as the input:
y_train_pred = tree_1.predict(X_train)

# Compute the residuals: The residuals are the differences between the predictions and the target column. 
# The predictions of X_train, defined here as y_train_pred, are subtracted from y_train, the target column, to compute the residuals:
y2_train = y_train - y_train_pred
# The residuals are defined as y2_train because they are the new target column for the next tree.

# Fit the new tree on the residuals: Fitting a new tree on the residuals is different than fitting a model on the training set. 
#The primary difference is in the predictions. In the bike rentals dataset, when fitting a new tree on the residuals, we should progressively get smaller numbers.
#Initialize a new tree and fit it on X_train and the residuals, y2_train:
tree_2 = DecisionTreeRegressor(max_depth=2, random_state=2)
tree_2.fit(X_train, y2_train)

# Repeat steps 2-4: As the process continues, the residuals should gradually approach 0 from the positive and negative direction. 
#The iterations continue for the number of estimators, n_estimators.
# Let's repeat the process for a third tree as follows:
y2_train_pred = tree_2.predict(X_train)
y3_train = y2_train - y2_train_pred
tree_3 = DecisionTreeRegressor(max_depth=2, random_state=2)
tree_3.fit(X_train, y3_train)

# As a learning purpose, stop the process here.
# Sum the results: Summing the results requires making predictions for each tree with the test set as follows:
y1_pred = tree_1.predict(X_test)
y2_pred = tree_2.predict(X_test)
y3_pred = tree_3.predict(X_test)

y_pred = y1_pred + y2_pred + y3_pred

# Lastly, let's compute the mean squared error (MSE) to obtain the results as follows:
from sklearn.metrics import mean_squared_error as MSE
MSE(y_test, y_pred)**0.5

#### Building a gradient boosting model in scikit-learn

Use GradientBoostingRegressor with hyperparameter adjustments. The advantage of using GradientBoostingRegressor is that it's much faster to build and easier to implement:

# import the regressor from the sklearn.ensemble library:
from sklearn.ensemble import GradientBoostingRegressor

# To obtain the same results, it's essential to match max_depth=2 and random_state=2. 
# Since there are only three trees, we must have n_estimators=3. 
# Finally, we must set the learning_rate=1.0 hyperparameter.
gbr = GradientBoostingRegressor(max_depth=2, n_estimators=3, random_state=2, learning_rate=1.0)

# the model has been initialized, it can be fit on the training data and scored against the test data:
gbr.fit(X_train, y_train)
y_pred = gbr.predict(X_test)
MSE(y_test, y_pred)**0.5

The result is the same to 11 decimal places.
Recall that the point of gradient boosting is to build a model with enough trees to transform a weak learner into a strong learner. This is easily done by changing n_estimators, the number of iterations, to a much larger number.

# Let's build and score a gradient boosting regressor with 30 estimators:
gbr = GradientBoostingRegressor(max_depth=2, n_estimators=30, random_state=2, learning_rate=1.0)
gbr.fit(X_train, y_train)
y_pred = gbr.predict(X_test)
MSE(y_test, y_pred)**0.5

# when n_estimators = 300:
gbr = GradientBoostingRegressor(max_depth=2, n_estimators=300, random_state=2, learning_rate=1.0)
gbr.fit(X_train, y_train)
y_pred = gbr.predict(X_test)
MSE(y_test, y_pred)**0.5

# remove parameter learning_rate:
gbr = GradientBoostingRegressor(max_depth=2, n_estimators=300, random_state=2)
gbr.fit(X_train, y_train)
y_pred = gbr.predict(X_test)
MSE(y_test, y_pred)**0.5

### Modifying GB Hyperparameters

learning_rate, also known as the shrinkage, shrinks the contribution of individual trees so that no tree has too much influence when building the model. If an entire ensemble is built from the errors of one base learner, without careful adjustment of hyperparameters, early trees in the model can have too much influence on subsequent development. learning_rate limits the influence of individual trees. Generally speaking, as n_estimators, the number of trees, goes up, learning_rate should go down.
learning_rate ranges from 0 to 1. A learning_rate value of 1 means that no adjustments are made. The default value of 0.1 means that the tree's influence is weighted at 10%.

# build and score a new GradientBoostingRegressor to see how the scores compare:
learning_rate_values = [0.001, 0.01, 0.05, 0.1, 0.15, 0.2, 0.3, 0.5, 1.0]

for value in learning_rate_values:
    gbr = GradientBoostingRegressor(max_depth=2,   n_estimators=300, random_state=2, learning_rate=value)
    gbr.fit(X_train, y_train)
    y_pred = gbr.predict(X_test)
    rmse = MSE(y_test, y_pred)**0.5
    print('Learning Rate:', value, ', Score:', rmse)

##### the default learning_rate value of 0.1 gives the best score for 300 trees.It's important to tune learning_rate and n_estimators together.

Base learner
The initial decision tree in the gradient boosting regressor is called the base learner because it's at the base of the ensemble. It's the first learner in the process. It is a weak learner transforming into a strong learner.

#It's possible to tune base learners for gains in accuracy.
# select a max_depth value of 1, 2, 3, or 4 and compare results as follows:
depths = [None, 1, 2, 3, 4]

for depth in depths:
    gbr = GradientBoostingRegressor(max_depth=depth, n_estimators=300, random_state=2)
    gbr.fit(X_train, y_train)
    y_pred = gbr.predict(X_test)
    rmse = MSE(y_test, y_pred)**0.5
    print('Max Depth:', depth, ', Score:', rmse)

subsample 
is a subset of samples. Since samples are the rows, a subset of rows means that all rows may not be included when building each tree. By changing subsample from 1.0 to a smaller decimal, trees only select that percentage of samples during the build phase. For example, subsample=0.8 would select 80% of samples for each tree.

samples = [1, 0.9, 0.8, 0.7, 0.6, 0.5]

for sample in samples:
    gbr = GradientBoostingRegressor(max_depth=3, n_estimators=300, subsample=sample, random_state=2)
    gbr.fit(X_train, y_train)
    y_pred = gbr.predict(X_test)
    rmse = MSE(y_test, y_pred)**0.5
    print('Subsample:', sample, ', Score:', rmse)

RandomizedSearchCV
Our preliminary analysis indicated that a grid search centered around max_depth=3, subsample=0.7, n_estimators=300, and learning_rate = 0.1 is a good place to start. As n_estimators goes up, learning_rate should go down..RandomizedSearchCV is preferred over GridSearchCV when there is too many combinations of hyperparameters. It will speed up computations.

# starting point:
params={'subsample':[0.65, 0.7, 0.75],
        'n_estimators':[300, 500, 1000],
         'learning_rate':[0.05, 0.075, 0.1]}

#import RandomizedSearchCV and initialize a gradient boosting model:
from sklearn.model_selection import RandomizedSearchCV

gbr = GradientBoostingRegressor(max_depth=3, random_state=2)

# Initialize RandomizedSearchCV with gbr and params as inputs in addition to the number of iterations, the scoring, and the number of folds. 
# Recall that n_jobs=-1 may speed up computations and random_state=2 ensures the consistency of results:
rand_reg = RandomizedSearchCV(gbr, params, n_iter=10, scoring='neg_mean_squared_error', cv=5, n_jobs=-1, random_state=2)

# fit the model on the training set and obtain the best parameters and scores:
from sklearn.metrics import mean_squared_error as MSE

rand_reg.fit(X_train, y_train)
best_model = rand_reg.best_estimator_
best_params = rand_reg.best_params_

print("Best params:", best_params)
best_score = np.sqrt(-rand_reg.best_score_)
print("Training score: {:.3f}".format(best_score))

y_pred = best_model.predict(X_test)
rmse_test = MSE(y_test, y_pred)**0.5
print('Test set score: {:.3f}'.format(rmse_test))

# after experimentations, a better model is possible:
gbr = GradientBoostingRegressor(max_depth=3, n_estimators=1600, subsample=0.75, learning_rate=0.02, random_state=2)
gbr.fit(X_train, y_train)
y_pred = gbr.predict(X_test)
MSE(y_test, y_pred)**0.5

XGBoost
XGBoost is an advanced version of gradient boosting with the same general structure, it transforms weak learners into strong learners by summing the residuals of trees.
The only difference in hyperparameters from the last section is that XGBoost refers to learning_rate as eta.

# intialize the model:
from xgboost import XGBRegressor

xg_reg = XGBRegressor(max_depth=3, n_estimators=1600, eta=0.02, subsample=0.75, random_state=2)
xg_reg.fit(X_train, y_train)
y_pred = xg_reg.predict(X_test)
MSE(y_test, y_pred)**0.5

##### XGBoost is preferred over gradient boosting in general because it consistently delivers better results, and because it's faster.

### Gradient boosting versus XGBoost - exoplanet example

We examine exoplanets over time. The dataset has 5,087 rows and 3,189 columns that record light flux at different times of a star's life cycle. Multiplying columns and rows together results in 1.5 million data points. Using a baseline of 100 trees, we need 150 million data points to build a model.
Each row is an individual star and the columns reveal different light patterns over time. In addition to light patterns, an exoplanet column is labeled 2 if the star hosts an exoplanet; otherwise, it is labeled 1.

# # Load the exoplanet dataset:
df = pd.read_csv('exoplanets.csv')
df.head()

# are all columns numerical?
df.info()

# number of null values?
df.isnull().sum().sum()

# split the data into training and test sets. 
#Note that the 0th column is the target column, y, and all other columns are the predictor columns, X:
X = df.iloc[:,1:]
y = df.iloc[:,0]
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=2)

#### Building gradient boosting classifiers

# import GradientBoostingClassifer and XGBClassifier in addition to accuracy_score so that we may compare both models:
from sklearn.ensemble import GradientBoostingClassifier

from xgboost import XGBClassifier

from sklearn.metrics import accuracy_score

#### Compare models using a timer

import time

# try an example:
start = time.time()
df.info()
end = time.time()
elapsed = end - start

print('\nRun Time: ' + str(elapsed) + ' seconds.')

####  Compare GradientBoostingClassifier and XGBoostClassifier with the exoplanet dataset for its speed

#GradientBoostingClassifier:
start = time.time()
gbr = GradientBoostingClassifier(n_estimators=100, max_depth=2, random_state=2)
gbr.fit(X_train, y_train)
y_pred = gbr.predict(X_test)
score = accuracy_score(y_pred, y_test)
print('Score: ' + str(score))
end = time.time()

elapsed = end - start
print('\nRun Time: ' + str(elapsed) + ' seconds')

# XGBClassifier:
start = time.time()
xg_reg = XGBClassifier(n_estimators=100, max_depth=2, random_state=2)
xg_reg.fit(X_train, y_train)
y_pred = xg_reg.predict(X_test)
score = accuracy_score(y_pred, y_test)

print('Score: ' + str(score))
end = time.time()

elapsed = end - start
print('Run Time: ' + str(elapsed) + ' seconds')
